% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{entropy}
\alias{entropy}
\title{Compute the entropy}
\usage{
entropy(pvector)
}
\arguments{
\item{pvector}{a vector of probability distribution}
}
\value{
von Neumann entropy
}
\description{
Compute the entropy
}

