% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{JSD}
\alias{JSD}
\title{Compute Jenson-Shannon divergence}
\usage{
JSD(p1vector, p2vector)
}
\arguments{
\item{p1vector}{a vector of probability distribution}

\item{p2vector}{a vector of probability distribution}
}
\value{
Jensen-Shannon divergence
}
\description{
Compute Jenson-Shannon divergence
}

