% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{KLD}
\alias{KLD}
\title{Compute Kullback-Leibler divergence}
\usage{
KLD(p1vector, p2vector)
}
\arguments{
\item{p1vector}{a vector of probability distribution}

\item{p2vector}{a vector of probability distribution}
}
\value{
Kullback-Leibler divergence
}
\description{
Compute Kullback-Leibler divergence
}

