% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{normalize.pvector}
\alias{normalize.pvector}
\title{Normalize probability distribution}
\usage{
normalize.pvector(pvector)
}
\arguments{
\item{pvector}{a vector of all probability distribution}
}
\value{
a norm vector for all probability distribution
}
\description{
Normalize probability distribution
}

